item sword

+++ GhostBattle.scala
@@ -11,13 +11,15 @@ class GhostBattle(actor: Player, ghost: Entity) extends Battle:
   def help: String = ""
 
   def attack: String =
-    val damageInflicted = Random.nextInt(34)
-    this.ghost.takeDamage(damageInflicted)
-    if this.ghost.health <= 0 then
-      this.ghost.pacify()
-      this.actor.endBattle()
-      "umm... you do know you cant kill ghosts, right?"
-    else s"Your attack deals ${damageInflicted} damage"
+    if actor.has(sword) then
+      val damageInflicted = Random.nextInt(34)
+      this.ghost.takeDamage(damageInflicted)
+      if this.ghost.health <= 0 then
+        this.ghost.pacify()
+        this.actor.endBattle()
+        "umm... you do know you cant kill ghosts, right?"
+      else s"Your attack deals ${damageInflicted} damage"
+    else "You can't expect to beat me without a weapon."
 
   def chat: String =
     if math.random < 0.25 then
